split args_catalog
