+++ src/epygram/cli/__init__.py
@@ -10,9 +10,8 @@ import sys
 import importlib
 import epygram
 
-
 commands = [f[:-3] for f in os.listdir(os.path.dirname(__file__))
-            if f.endswith('.py') and f not in ('__init__.py', 'args_catalog.py')]
+            if f.endswith('.py') and f not in ('__init__.py',)]
 
 
 def main():
