+++ src/epygram/cli/histogram.py
@@ -13,9 +13,9 @@ from bronx.syntax.pretty import smooth_string
 import epygram
 from epygram import epylog, epygramError
 from .args_catalog import (add_arg_to_parser,
-                           files_management, fields_management,
-                           misc_options, output_options,
-                           runtime_options, graphical_options)
+                           files_args, fields_args,
+                           misc_args, output_args,
+                           runtime_args, graphical_args)
 
 import matplotlib.pyplot as plt
 
@@ -265,29 +265,29 @@ def get_args():
     parser = argparse.ArgumentParser(description='An EPyGrAM tool for making histograms \
                                                   of meteorological fields from a resource.',
                                      epilog='End of help for: %(prog)s (EPyGrAM-' + epygram.__version__ + ')')
-    add_arg_to_parser(parser, files_management['principal_file'])
-    add_arg_to_parser(parser, fields_management['field'])
-    add_arg_to_parser(parser, fields_management['windfieldU'])
-    add_arg_to_parser(parser, fields_management['windfieldV'])
+    add_arg_to_parser(parser, files_args['principal_file'])
+    add_arg_to_parser(parser, fields_args['field'])
+    add_arg_to_parser(parser, fields_args['windfieldU'])
+    add_arg_to_parser(parser, fields_args['windfieldV'])
     diffmodes = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(diffmodes, files_management['file_to_refer_in_diff'])
-    add_arg_to_parser(diffmodes, files_management['file_to_refer_in_diffonly'])
-    add_arg_to_parser(parser, output_options['output'])
-    add_arg_to_parser(parser, output_options['outputfilename'])
-    add_arg_to_parser(parser, misc_options['LAMzone'])
-    add_arg_to_parser(parser, graphical_options['bins'])
-    add_arg_to_parser(parser, graphical_options['diffbins'])
-    add_arg_to_parser(parser, graphical_options['minmax'])
-    add_arg_to_parser(parser, graphical_options['diffminmax'])
-    add_arg_to_parser(parser, graphical_options['center_hist_on_0'])
-    add_arg_to_parser(parser, graphical_options['diff_center_hist_on_0'])
-    add_arg_to_parser(parser, graphical_options['legend'])
-    add_arg_to_parser(parser, graphical_options['lonlat_zoom'])
-    add_arg_to_parser(parser, graphical_options['figures_dpi'])
-    add_arg_to_parser(parser, misc_options['operation_on_field'])
-    add_arg_to_parser(parser, misc_options['diffoperation_on_field'])
-    add_arg_to_parser(parser, misc_options['mask_threshold'])
-    add_arg_to_parser(parser, runtime_options['verbose'])
+    add_arg_to_parser(diffmodes, files_args['file_to_refer_in_diff'])
+    add_arg_to_parser(diffmodes, files_args['file_to_refer_in_diffonly'])
+    add_arg_to_parser(parser, output_args['output'])
+    add_arg_to_parser(parser, output_args['outputfilename'])
+    add_arg_to_parser(parser, misc_args['LAMzone'])
+    add_arg_to_parser(parser, graphical_args['bins'])
+    add_arg_to_parser(parser, graphical_args['diffbins'])
+    add_arg_to_parser(parser, graphical_args['minmax'])
+    add_arg_to_parser(parser, graphical_args['diffminmax'])
+    add_arg_to_parser(parser, graphical_args['center_hist_on_0'])
+    add_arg_to_parser(parser, graphical_args['diff_center_hist_on_0'])
+    add_arg_to_parser(parser, graphical_args['legend'])
+    add_arg_to_parser(parser, graphical_args['lonlat_zoom'])
+    add_arg_to_parser(parser, graphical_args['figures_dpi'])
+    add_arg_to_parser(parser, misc_args['operation_on_field'])
+    add_arg_to_parser(parser, misc_args['diffoperation_on_field'])
+    add_arg_to_parser(parser, misc_args['mask_threshold'])
+    add_arg_to_parser(parser, runtime_args['verbose'])
     args = parser.parse_args()
 
     # 2. Initializations
