+++ src/epygram/cli/movefield.py
@@ -12,8 +12,8 @@ import epygram
 from epygram import epylog, epygramError
 from epygram.extra import griberies
 from .args_catalog import (add_arg_to_parser,
-                           files_management, fields_management,
-                           runtime_options)
+                           files_args, fields_args,
+                           runtime_args)
 
 
 def main():
@@ -149,20 +149,20 @@ def get_args():
     parser = argparse.ArgumentParser(description="An EPyGrAM tool for moving field(s) from a resource to another\
                                                   (fields are not removed from the source file).",
                                      epilog='End of help for: %(prog)s (EPyGrAM-' + epygram.__version__ + ')')
-    add_arg_to_parser(parser, files_management['principal_file'])
-    add_arg_to_parser(parser, files_management['source_file'])
-    add_arg_to_parser(parser, files_management['in_place'])
+    add_arg_to_parser(parser, files_args['principal_file'])
+    add_arg_to_parser(parser, files_args['source_file'])
+    add_arg_to_parser(parser, files_args['in_place'])
     flds = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(flds, fields_management['field'])
-    add_arg_to_parser(flds, fields_management['list_of_fields'])
+    add_arg_to_parser(flds, fields_args['field'])
+    add_arg_to_parser(flds, fields_args['list_of_fields'])
     diff = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(diff, files_management['replace_by_diff'])
-    add_arg_to_parser(diff, files_management['replace_by_reversediff'])
-    add_arg_to_parser(diff, files_management['replace_by_addition'])
-    add_arg_to_parser(diff, files_management['replace_by_product'])
+    add_arg_to_parser(diff, files_args['replace_by_diff'])
+    add_arg_to_parser(diff, files_args['replace_by_reversediff'])
+    add_arg_to_parser(diff, files_args['replace_by_addition'])
+    add_arg_to_parser(diff, files_args['replace_by_product'])
     status = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(status, runtime_options['verbose'])
-    add_arg_to_parser(status, runtime_options['percentage'])
+    add_arg_to_parser(status, runtime_args['verbose'])
+    add_arg_to_parser(status, runtime_args['percentage'])
     args = parser.parse_args()
 
     # 2. Initializations
