+++ src/epygram/cli/plot3d.py
@@ -14,10 +14,10 @@ from footprints import proxy as fpx
 import epygram
 from epygram import epylog, epygramError
 from .args_catalog import (add_arg_to_parser,
-                           files_management, fields_management,
-                           misc_options, output_options,
-                           runtime_options, graphical_options,
-                           extraction_options)
+                           files_args, fields_args,
+                           misc_args, output_args,
+                           runtime_args, graphical_args,
+                           extraction_args)
 from epygram.base import FieldSet
 from epygram.geometries import VGeometry
 
@@ -649,54 +649,54 @@ def get_args():
                                                            'Uses and requires the Python package `vtk`.']),
                                      epilog='End of help for: %(prog)s (EPyGrAM-' + epygram.__version__ + ')')
 
-    add_arg_to_parser(parser, graphical_options['plotmode'])
-    add_arg_to_parser(parser, files_management['principal_file'])
-    add_arg_to_parser(parser, fields_management['field'])
-    add_arg_to_parser(parser, fields_management['windfieldU'])
-    add_arg_to_parser(parser, fields_management['windfieldV'])
+    add_arg_to_parser(parser, graphical_args['plotmode'])
+    add_arg_to_parser(parser, files_args['principal_file'])
+    add_arg_to_parser(parser, fields_args['field'])
+    add_arg_to_parser(parser, fields_args['windfieldU'])
+    add_arg_to_parser(parser, fields_args['windfieldV'])
     diffmodes = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(diffmodes, files_management['file_to_refer_in_diff'])
-    add_arg_to_parser(diffmodes, files_management['file_to_refer_in_diffonly'])
-    add_arg_to_parser(parser, output_options['output'])
-    add_arg_to_parser(parser, output_options['outputfilename'])
-    add_arg_to_parser(parser, misc_options['LAMzone'])
-    add_arg_to_parser(parser, graphical_options['specific_map_projection'])
-    add_arg_to_parser(parser, graphical_options['minmax'])
-    add_arg_to_parser(parser, graphical_options['diffminmax'])
-    add_arg_to_parser(parser, graphical_options['levels_number'])
-    add_arg_to_parser(parser, graphical_options['diff_levels_number'])
-    add_arg_to_parser(parser, graphical_options['legend'])
-    add_arg_to_parser(parser, graphical_options['hide_axes'])
-    add_arg_to_parser(parser, graphical_options['vectors_subsampling'])
-    add_arg_to_parser(parser, graphical_options['vectors_verticalsubsampling'])
-    add_arg_to_parser(parser, graphical_options['vectors_scale_factor'])
-    add_arg_to_parser(parser, graphical_options['diffvectors_scale_factor'])
-    add_arg_to_parser(parser, graphical_options['streamlines_time'])
-    add_arg_to_parser(parser, graphical_options['diffstreamlines_time'])
-    add_arg_to_parser(parser, graphical_options['colorminmax'])
-    add_arg_to_parser(parser, graphical_options['diffcolorminmax'])
-    add_arg_to_parser(parser, graphical_options['alpha'])
-    add_arg_to_parser(parser, graphical_options['alphaminmax'])
-    add_arg_to_parser(parser, graphical_options['diffalphaminmax'])
-    add_arg_to_parser(parser, graphical_options['lonlat_zoom'])
-    add_arg_to_parser(parser, graphical_options['z_factor'])
-    add_arg_to_parser(parser, graphical_options['resolution_increase'])
-    add_arg_to_parser(parser, graphical_options['window_size'])
-    add_arg_to_parser(parser, graphical_options['ground'])
-    add_arg_to_parser(parser, graphical_options['background_color'])
-    add_arg_to_parser(parser, graphical_options['global_shift_center'])
-    add_arg_to_parser(parser, graphical_options['focal_point'])
-    add_arg_to_parser(parser, graphical_options['camera'])
+    add_arg_to_parser(diffmodes, files_args['file_to_refer_in_diff'])
+    add_arg_to_parser(diffmodes, files_args['file_to_refer_in_diffonly'])
+    add_arg_to_parser(parser, output_args['output'])
+    add_arg_to_parser(parser, output_args['outputfilename'])
+    add_arg_to_parser(parser, misc_args['LAMzone'])
+    add_arg_to_parser(parser, graphical_args['specific_map_projection'])
+    add_arg_to_parser(parser, graphical_args['minmax'])
+    add_arg_to_parser(parser, graphical_args['diffminmax'])
+    add_arg_to_parser(parser, graphical_args['levels_number'])
+    add_arg_to_parser(parser, graphical_args['diff_levels_number'])
+    add_arg_to_parser(parser, graphical_args['legend'])
+    add_arg_to_parser(parser, graphical_args['hide_axes'])
+    add_arg_to_parser(parser, graphical_args['vectors_subsampling'])
+    add_arg_to_parser(parser, graphical_args['vectors_verticalsubsampling'])
+    add_arg_to_parser(parser, graphical_args['vectors_scale_factor'])
+    add_arg_to_parser(parser, graphical_args['diffvectors_scale_factor'])
+    add_arg_to_parser(parser, graphical_args['streamlines_time'])
+    add_arg_to_parser(parser, graphical_args['diffstreamlines_time'])
+    add_arg_to_parser(parser, graphical_args['colorminmax'])
+    add_arg_to_parser(parser, graphical_args['diffcolorminmax'])
+    add_arg_to_parser(parser, graphical_args['alpha'])
+    add_arg_to_parser(parser, graphical_args['alphaminmax'])
+    add_arg_to_parser(parser, graphical_args['diffalphaminmax'])
+    add_arg_to_parser(parser, graphical_args['lonlat_zoom'])
+    add_arg_to_parser(parser, graphical_args['z_factor'])
+    add_arg_to_parser(parser, graphical_args['resolution_increase'])
+    add_arg_to_parser(parser, graphical_args['window_size'])
+    add_arg_to_parser(parser, graphical_args['ground'])
+    add_arg_to_parser(parser, graphical_args['background_color'])
+    add_arg_to_parser(parser, graphical_args['global_shift_center'])
+    add_arg_to_parser(parser, graphical_args['focal_point'])
+    add_arg_to_parser(parser, graphical_args['camera'])
     Y = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(Y, extraction_options['verticalcoord2pressure'])
-    add_arg_to_parser(Y, extraction_options['verticalcoord2height'])
-    add_arg_to_parser(Y, extraction_options['verticalcoord2altitude'])
-    add_arg_to_parser(parser, extraction_options['no_cheap_height_conversion'])
-    add_arg_to_parser(parser, misc_options['pressure_unit_hpa'])
-    add_arg_to_parser(parser, misc_options['operation_on_field'])
-    add_arg_to_parser(parser, misc_options['diffoperation_on_field'])
-    add_arg_to_parser(parser, misc_options['composition_with_field'])
-    add_arg_to_parser(parser, runtime_options['verbose'])
+    add_arg_to_parser(Y, extraction_args['verticalcoord2pressure'])
+    add_arg_to_parser(Y, extraction_args['verticalcoord2height'])
+    add_arg_to_parser(Y, extraction_args['verticalcoord2altitude'])
+    add_arg_to_parser(parser, extraction_args['no_cheap_height_conversion'])
+    add_arg_to_parser(parser, misc_args['pressure_unit_hpa'])
+    add_arg_to_parser(parser, misc_args['operation_on_field'])
+    add_arg_to_parser(parser, misc_args['diffoperation_on_field'])
+    add_arg_to_parser(parser, misc_args['composition_with_field'])
+    add_arg_to_parser(parser, runtime_args['verbose'])
 
     args = parser.parse_args()
