+++ src/epygram/cli/profile.py
@@ -13,10 +13,10 @@ import epygram
 from epygram import epylog
 from epygram.fields.V1DField import plotprofiles
 from .args_catalog import (add_arg_to_parser,
-                           files_management, fields_management,
-                           misc_options, output_options,
-                           runtime_options, graphical_options,
-                           extraction_options)
+                           files_args, fields_args,
+                           misc_args, output_args,
+                           runtime_args, graphical_args,
+                           extraction_args)
 
 import matplotlib.pyplot as plt
 
@@ -297,34 +297,34 @@ def get_args():
     parser = argparse.ArgumentParser(description='An EPyGrAM tool for extracting (and plotting) vertical profiles\
                                                   of meteorological fields from a resource.',
                                      epilog='End of help for: %(prog)s (EPyGrAM-' + epygram.__version__ + ')')
-    add_arg_to_parser(parser, files_management['principal_file'])
-    add_arg_to_parser(parser, fields_management['vertical_field'])
-    add_arg_to_parser(parser, extraction_options['point_coordinates'],
+    add_arg_to_parser(parser, files_args['principal_file'])
+    add_arg_to_parser(parser, fields_args['vertical_field'])
+    add_arg_to_parser(parser, extraction_args['point_coordinates'],
                       required=False, default=None,
-                      help=(extraction_options['point_coordinates'][-1]['help'] +
+                      help=(extraction_args['point_coordinates'][-1]['help'] +
                             'If not given, make horizontally-averaged profile.'))
-    add_arg_to_parser(parser, extraction_options['horizontal_interpolation'])
-    add_arg_to_parser(parser, extraction_options['external_distance'])
+    add_arg_to_parser(parser, extraction_args['horizontal_interpolation'])
+    add_arg_to_parser(parser, extraction_args['external_distance'])
     diffmodes = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(diffmodes, files_management['file_to_refer_in_diff'])
-    add_arg_to_parser(diffmodes, files_management['file_to_refer_in_diffonly'])
-    add_arg_to_parser(parser, output_options['output'])
-    add_arg_to_parser(parser, output_options['outputfilename'])
-    add_arg_to_parser(parser, output_options['noplot'])
+    add_arg_to_parser(diffmodes, files_args['file_to_refer_in_diff'])
+    add_arg_to_parser(diffmodes, files_args['file_to_refer_in_diffonly'])
+    add_arg_to_parser(parser, output_args['output'])
+    add_arg_to_parser(parser, output_args['outputfilename'])
+    add_arg_to_parser(parser, output_args['noplot'])
     Y = parser.add_mutually_exclusive_group()
-    add_arg_to_parser(Y, extraction_options['verticalcoord2pressure'])
-    add_arg_to_parser(Y, extraction_options['verticalcoord2height'])
-    add_arg_to_parser(Y, extraction_options['verticalcoord2altitude'])
-    add_arg_to_parser(parser, extraction_options['no_cheap_height_conversion'])
-    add_arg_to_parser(parser, graphical_options['legend'])
-    add_arg_to_parser(parser, graphical_options['emagram_like_profiles'])
-    add_arg_to_parser(parser, graphical_options['vertical_logscale'])
-    add_arg_to_parser(parser, graphical_options['scientifical_unit'])
-    add_arg_to_parser(parser, graphical_options['vertical_zoom'])
-    add_arg_to_parser(parser, graphical_options['figures_dpi'])
-    add_arg_to_parser(parser, misc_options['operation_on_field'])
-    add_arg_to_parser(parser, misc_options['diffoperation_on_field'])
-    add_arg_to_parser(parser, runtime_options['verbose'])
+    add_arg_to_parser(Y, extraction_args['verticalcoord2pressure'])
+    add_arg_to_parser(Y, extraction_args['verticalcoord2height'])
+    add_arg_to_parser(Y, extraction_args['verticalcoord2altitude'])
+    add_arg_to_parser(parser, extraction_args['no_cheap_height_conversion'])
+    add_arg_to_parser(parser, graphical_args['legend'])
+    add_arg_to_parser(parser, graphical_args['emagram_like_profiles'])
+    add_arg_to_parser(parser, graphical_args['vertical_logscale'])
+    add_arg_to_parser(parser, graphical_args['scientifical_unit'])
+    add_arg_to_parser(parser, graphical_args['vertical_zoom'])
+    add_arg_to_parser(parser, graphical_args['figures_dpi'])
+    add_arg_to_parser(parser, misc_args['operation_on_field'])
+    add_arg_to_parser(parser, misc_args['diffoperation_on_field'])
+    add_arg_to_parser(parser, runtime_args['verbose'])
     args = parser.parse_args()
 
     # 2. Initializations
